corrected user responses

+++ src/modules/user/user.controller.ts
@@ -107,19 +107,50 @@ const updateUser = async (req: Request, res: Response) => {
 const deleteUser = async (req: Request, res: Response) => {
   // console.log(req.params.id);
   try {
-    const result = await userServices.deleteUser(req.params.userId!);
-
-    if (result.rowCount === 0) {
-      res.status(404).json({
-        success: false,
-        message: "User not found",
-      });
-    } else {
-      res.status(200).json({
-        success: true,
-        message: "User deleted successfully",
-        data: result.rows,
-      });
+    if (req.user?.role === "admin") {
+      const userBookingData = await userServices.getAllBookingsByUser(
+        req.params.userId!
+      );
+      let activeBookings = 0;
+      if (userBookingData?.rows.length > 0) {
+        userBookingData.rows.forEach((booking: any) => {
+          if (booking.status === "active") {
+            activeBookings += 1;
+          }
+        });
+        if (activeBookings > 0) {
+          return res.status(400).json({
+            success: false,
+            message: "Cannot delete user with active bookings",
+          });
+        } else {
+          const result = await userServices.deleteUser(req.params.userId!);
+          if (result.rowCount === 0) {
+            res.status(404).json({
+              success: false,
+              message: "User not found",
+            });
+          } else {
+            res.status(200).json({
+              success: true,
+              message: "User deleted successfully",
+            });
+          }
+        }
+      } else {
+        const result = await userServices.deleteUser(req.params.userId!);
+        if (result.rowCount === 0) {
+          res.status(404).json({
+            success: false,
+            message: "User not found",
+          });
+        } else {
+          res.status(200).json({
+            success: true,
+            message: "User deleted successfully",
+          });
+        }
+      }
     }
   } catch (err: any) {
     res.status(500).json({
